Simple Account
Here is a simple implementation of a users account for preexisting and new. New account issues: wrong value for input to input comparing for passwords. Preexisting: just need to connect to vector to see if values are held in vector. I will work on it some more this was just to have a starting point.